+++ database/sampleData.js
@@ -8,7 +8,97 @@ var homes = [
   'https://s3-us-west-1.amazonaws.com/mashbnb/airbnb-georgia.jpg',
   'https://s3-us-west-1.amazonaws.com/mashbnb/full-room_orig.jpg',
   'https://s3-us-west-1.amazonaws.com/mashbnb/dunkinhouse.jpg',
-  'https://s3-us-west-1.amazonaws.com/mashbnb/Airbnb+Alaska.png'
+  'https://s3-us-west-1.amazonaws.com/mashbnb/Airbnb+Alaska.png',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/tiny-house-for-sale-nashville-tn-airbnb-cute-button.jpg',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/tiny-house-2-of-7.jpg',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/the-wildflower-airbnb-1519016646.jpg',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/thatch-cottage-airbnb-4+copy.jpg',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/Hero_airbnb_4.jpg',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/Romantic-AirBNBs-in-Paris-Airbnb-near-the-Louvre.png',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/Impeccable-Cute-Condo.jpg',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/IMG_20160817_145016-1024x576+copy.jpg',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/Schermafbeelding-2017-10-26-om-15.22.05+copy.png',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/best-bed-and-breakfasts-in-nyc-from-brownstones-to-studios-elegant-air-b-new-york-magnificent-6+copy.jpg',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/bird-island.jpg',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/origin-13897.jpg',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/full-room_orig.jpg',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/cute-house-in-lake-arrowhead-airbnb-property-2.jpg',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/sunny-living-room-large-sofa-airbnb-apartment-porto+copy.jpg',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/007235ea-aaaf-4dc0-b733-845b92ed5509.jpg',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/0_F43HRv7o6nBpfaWF.jpeg',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/12477131-airbnb-paris.jpg',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/63b5a1b7_original.jpg',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/AD-Cute-Tiny-Houses-In-Every-Single-State-06.png',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/airbnb-apartment-porto-living-room.jpg',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/airbnb-georgia.jpg',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/full-room_orig.jpg',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/dunkinhouse.jpg',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/Airbnb+Alaska.png',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/tiny-house-for-sale-nashville-tn-airbnb-cute-button.jpg',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/tiny-house-2-of-7.jpg',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/the-wildflower-airbnb-1519016646.jpg',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/thatch-cottage-airbnb-4+copy.jpg',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/Hero_airbnb_4.jpg',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/Romantic-AirBNBs-in-Paris-Airbnb-near-the-Louvre.png',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/Impeccable-Cute-Condo.jpg',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/IMG_20160817_145016-1024x576+copy.jpg',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/Schermafbeelding-2017-10-26-om-15.22.05+copy.png',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/best-bed-and-breakfasts-in-nyc-from-brownstones-to-studios-elegant-air-b-new-york-magnificent-6+copy.jpg',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/bird-island.jpg',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/origin-13897.jpg',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/full-room_orig.jpg',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/cute-house-in-lake-arrowhead-airbnb-property-2.jpg',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/sunny-living-room-large-sofa-airbnb-apartment-porto+copy.jpg',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/007235ea-aaaf-4dc0-b733-845b92ed5509.jpg',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/0_F43HRv7o6nBpfaWF.jpeg',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/12477131-airbnb-paris.jpg',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/63b5a1b7_original.jpg',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/AD-Cute-Tiny-Houses-In-Every-Single-State-06.png',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/airbnb-apartment-porto-living-room.jpg',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/airbnb-georgia.jpg',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/full-room_orig.jpg',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/dunkinhouse.jpg',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/Airbnb+Alaska.png',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/tiny-house-for-sale-nashville-tn-airbnb-cute-button.jpg',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/tiny-house-2-of-7.jpg',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/the-wildflower-airbnb-1519016646.jpg',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/thatch-cottage-airbnb-4+copy.jpg',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/Hero_airbnb_4.jpg',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/Romantic-AirBNBs-in-Paris-Airbnb-near-the-Louvre.png',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/Impeccable-Cute-Condo.jpg',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/IMG_20160817_145016-1024x576+copy.jpg',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/Schermafbeelding-2017-10-26-om-15.22.05+copy.png',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/best-bed-and-breakfasts-in-nyc-from-brownstones-to-studios-elegant-air-b-new-york-magnificent-6+copy.jpg',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/bird-island.jpg',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/origin-13897.jpg',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/full-room_orig.jpg',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/cute-house-in-lake-arrowhead-airbnb-property-2.jpg',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/sunny-living-room-large-sofa-airbnb-apartment-porto+copy.jpg',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/007235ea-aaaf-4dc0-b733-845b92ed5509.jpg',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/0_F43HRv7o6nBpfaWF.jpeg',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/12477131-airbnb-paris.jpg',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/63b5a1b7_original.jpg',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/AD-Cute-Tiny-Houses-In-Every-Single-State-06.png',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/airbnb-apartment-porto-living-room.jpg',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/airbnb-georgia.jpg',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/full-room_orig.jpg',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/dunkinhouse.jpg',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/Airbnb+Alaska.png',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/tiny-house-for-sale-nashville-tn-airbnb-cute-button.jpg',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/tiny-house-2-of-7.jpg',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/the-wildflower-airbnb-1519016646.jpg',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/thatch-cottage-airbnb-4+copy.jpg',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/Hero_airbnb_4.jpg',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/Romantic-AirBNBs-in-Paris-Airbnb-near-the-Louvre.png',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/Impeccable-Cute-Condo.jpg',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/IMG_20160817_145016-1024x576+copy.jpg',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/Schermafbeelding-2017-10-26-om-15.22.05+copy.png',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/best-bed-and-breakfasts-in-nyc-from-brownstones-to-studios-elegant-air-b-new-york-magnificent-6+copy.jpg',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/bird-island.jpg',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/origin-13897.jpg',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/full-room_orig.jpg',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/cute-house-in-lake-arrowhead-airbnb-property-2.jpg',
+  'https://s3-us-west-1.amazonaws.com/mashbnb/sunny-living-room-large-sofa-airbnb-apartment-porto+copy.jpg',
 ];
 
 module.exports = {
